Support for custom layout in Nlog target
Nlog supports a variety of layout renderers (see. https://github.com/nlog/nlog/wiki/Layout-Renderers)
which can be used via the "layout" attribute in the "target" element in Web.config.
Supporting this attribute relies on inheriting from TargetWithLayout and rendering the message
using the layout, if the latter is overridden.